Project_2/network-monitor/requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
PyYAML>=6.0.1
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
